+++ src/client/components/MainMenu.js
@@ -2,7 +2,6 @@ import React, { Component } from "react";
 import SOButton from './SOButton';
 import JoinGame from './JoinGame';
 import NewGame from './NewGame';
-import Connection from '../utils/Connection'
 
 export default class MainMenu extends Component {
   constructor(props) {
@@ -27,124 +26,8 @@ export default class MainMenu extends Component {
     };
 
     var self = this;
-    
-    let server = new Connection(function(data) {
-      var round = data.round;
-      var me = data.you;
 
-      //determine if the server is waiting on us
-      var isWaiting = (function() {
-        for (var i = 0; i < round.waitingList.length; i++) {
-          if (me.id === round.waitingList[i].id) {
-            return true;
-          }
-        }
-        return false;
-      })();
-
-      data.message = "";
-      switch (round.phase) {
-        case "start":
-          if (isWaiting) {
-            self.props.changePage(StartPage, data);
-          } else {
-            data.message = "Waiting to begin the game...";
-            self.props.changePage(Waiting, data);
-          }
-          break;
-        case "selection":
-          if (me.isCaptain) {
-            self.props.changePage(SelectionPhase, data);
-          } else {
-            //find out who's captain
-            for (var i = 0; i < round.players.length; i++) {
-              if (round.players[i].isCaptain) {
-                data.message +=
-                  round.players[i].name + " is selecting players for ";
-                break;
-              }
-            }
-            data.message += "mission " + round.currentMission.number + "...";
-            self.props.changePage(CaptainWaiting, data);
-          }
-
-          break;
-        case "voting":
-          //determine if we have already voted
-          var alreadyVoted = false;
-          var votes = round.currentMission.votes;
-          var me = data.you;
-          for (var i = 0; i < votes.length; i++) {
-            if (votes[i].playerId === me.id) {
-              alreadyVoted = true;
-              break;
-            }
-          }
-
-          if (!alreadyVoted) {
-            self.props.changePage(VotingPhase, data);
-          } else {
-            data.message = "Waiting for players to vote...";
-            self.props.changePage(Waiting, data);
-          }
-          break;
-        case "voting_results":
-          if (isWaiting) {
-            self.props.changePage(VotingResults, data);
-          } else {
-            data.message =
-              "Waiting for everyone to finish viewing the results...";
-            self.props.changePage(Waiting, data);
-          }
-          break;
-        case "mission":
-          //figure out if we are on this mission
-          var isOnMission = false;
-          for (
-            var i = 0;
-            i < round.currentMission.playersOnMission.length;
-            i++
-          ) {
-            var thisPlayer = round.currentMission.playersOnMission[i];
-            if (thisPlayer.id === me.id) {
-              isOnMission = true;
-              break;
-            }
-          }
-
-          //if we are on the mission, figure out if we've voted already
-          var hasntVotedYet;
-          var missionVotes = round.currentMission.missionVotes;
-          if (isOnMission) {
-            hasntVotedYet = true;
-            for (var i = 0; i < missionVotes.length; i++) {
-              if (missionVotes[i].playerId === me.id) {
-                hasntVotedYet = false;
-                break;
-              }
-            }
-          }
-
-          if (isOnMission && hasntVotedYet) {
-            self.props.changePage(MissionPhase, data);
-          } else {
-            data.message = "Waiting for the mission to complete...";
-            self.props.changePage(Waiting, data);
-          }
-
-          break;
-        case "mission_results":
-          if (isWaiting) {
-            self.props.changePage(Results, data);
-          } else {
-            data.message =
-              "Waiting for everyone to finish viewing the results...";
-            self.props.changePage(Waiting, data);
-          }
-          break;
-      }
-    });
-    server.on("joinGame", function(data) {
+    this.props.server.on("joinGame", function(data) {
       if (data.success) {
         self.props.changePage(Lobby, data.game);
         server.on("updatePlayerList", function(data) {
@@ -154,7 +37,8 @@ export default class MainMenu extends Component {
         alert("Failed to join game!");
       }
     });
-    server.on("replace", function(data) {
+    
+    this.props.server.on("replace", function(data) {
       self.props.changePage(Replace, data);
     });
